azure_openai_func_calling_without_mcp/host/requirements.txt
requests
rich
openai>=1.3.5
python-dotenv